Modifications après que monsieur Remy ait dit que unsigned int long c'était pareil que unsigned int, voilà voilà

+++ face.cpp
@@ -7,9 +7,9 @@ Face::Face() {}
 Face::~Face() {}
 
 //Getters sommets
-long unsigned Face::getS_one() const { return summit_one; }
-long unsigned Face::getS_two() const { return summit_two; }
-long unsigned Face::getS_three() const { return summit_three; }
+unsigned int Face::getS_one() const { return summit_one; }
+unsigned int Face::getS_two() const { return summit_two; }
+unsigned int Face::getS_three() const { return summit_three; }
 //Getters segments
 double Face::getSeg_one() const { return segment_one; }
 double Face::getSeg_two() const { return segment_two; }
@@ -20,9 +20,9 @@ double Face::getPerimeter() const { return perimeter; }
 double Face::getArea() const { return area; }
 
 //Setteur sommets
-void Face::setS_one(const long unsigned& _one) { summit_one = _one; }
-void Face::setS_two(const long unsigned& _two) { summit_two = _two; }
-void Face::setS_three(const long unsigned& _three) { summit_three = _three; }
+void Face::setS_one(const unsigned int& _one) { summit_one = _one; }
+void Face::setS_two(const unsigned int& _two) { summit_two = _two; }
+void Face::setS_three(const unsigned int& _three) { summit_three = _three; }
 //Setteur segments
 void Face::setSeg_one(const double& s_one) { segment_one = s_one; }
 void Face::setSeg_two(const double& s_two) { segment_two = s_two; }
